Revert "Reapply "temp""
This reverts commit 108a3b1889dce19eff22d6198a79d2f321113e6c.

+++ src/playground_base/playground_base/base/topo/param/FBBoolParam.hpp
@@ -7,15 +7,15 @@ struct FBBoolParam
 {
   int ValueCount() const { return 2; }
 
-  double PlainToNormalized(bool plain) const;
-  bool NormalizedToPlain(double normalized) const;
+  float PlainToNormalized(bool plain) const;
+  bool NormalizedToPlain(float normalized) const;
 
   std::string PlainToText(bool plain) const;
   std::optional<bool> TextToPlain(std::string const& text) const;
 };
 
 inline bool
-FBBoolParam::NormalizedToPlain(double normalized) const
+FBBoolParam::NormalizedToPlain(float normalized) const
 {
   return normalized >= 0.5f;
 }
